Restructuring JS_DROP_FILE and adding comments to test.py
